Split overview, dynamic, static & prepared stmts
