+++ test/snippets/include/snippets/credentials.hpp
@@ -1,0 +1,23 @@
+//
+// Copyright (c) 2019-2024 Ruben Perez Hidalgo (rubenperez038 at gmail dot com)
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+//
+
+#ifndef BOOST_MYSQL_TEST_SNIPPETS_INCLUDE_SNIPPETS_CREDENTIALS_HPP
+#define BOOST_MYSQL_TEST_SNIPPETS_INCLUDE_SNIPPETS_CREDENTIALS_HPP
+
+namespace boost {
+namespace mysql {
+namespace test {
+
+constexpr const char* mysql_username = "example_user";
+constexpr const char* mysql_password = "example_password";
+constexpr const char* example_db = "boost_mysql_examples";
+
+}  // namespace test
+}  // namespace mysql
+}  // namespace boost
+
+#endif
