+++ test/snippets/snippets.cpp
@@ -129,8 +129,6 @@ void run_coro(boost::asio::any_io_executor ex, std::function<boost::asio::awaita
 }
 #endif
 
-const char* get_value_from_user() { return ""; }
-int get_int_value_from_user() { return 42; }
 std::int64_t get_employee_id() { return 42; }
 std::string get_company_id() { return "HGS"; }
 
@@ -149,28 +147,6 @@ struct post
 // We must use Boost.Describe to add reflection capabilities to post.
 // We must list all the fields that should be populated by Boost.MySQL
 BOOST_DESCRIBE_STRUCT(post, (), (id, title, body))
-//]
-
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-//[describe_post_v2
-struct post_v2
-{
-    int id;
-    std::string title;
-    std::optional<std::string> body;  // body may be NULL
-};
-BOOST_DESCRIBE_STRUCT(post_v2, (), (id, title, body))
-//]
-#endif
-
-//[describe_statistics
-struct statistics
-{
-    std::string company;
-    double average;
-    double max_value;
-};
-BOOST_DESCRIBE_STRUCT(statistics, (), (company, average, max_value))
 //]
 
 //[describe_stored_procedures
@@ -196,540 +172,6 @@ BOOST_DESCRIBE_STRUCT(employee, (), (first_name, last_name, salary))
 // We can use an empty tuple to represent it.
 using empty = std::tuple<>;
 //]
-
-//[prepared_statements_execute
-// description, price and show_in_store are not trusted, since they may
-// have been read from a file or an HTTP endpoint
-void insert_product(
-    tcp_ssl_connection& conn,
-    const statement& stmt,
-    string_view description,
-    int price,
-    bool show_in_store
-)
-{
-    results result;
-    conn.execute(stmt.bind(description, price, show_in_store), result);
-}
-//]
-
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-//[prepared_statements_execute_null
-// description, price and show_in_store are not trusted, since they may
-// have been read from a file or an HTTP endpoint
-void insert_product(
-    tcp_ssl_connection& conn,
-    const statement& stmt,
-    std::optional<string_view> description,
-    int price,
-    bool show_in_store
-)
-{
-    // If description has a value, a string will be sent to the server; otherwise, a NULL will
-    results result;
-    conn.execute(stmt.bind(description, price, show_in_store), result);
-}
-//]
-void run_insert_product_optional(tcp_ssl_connection& conn, const statement& stmt)
-{
-    insert_product(conn, stmt, std::optional<string_view>(), 2000, true);
-}
-#else
-void run_insert_product_optional(tcp_ssl_connection&, const statement&) {}
-#endif
-
-//[prepared_statements_execute_iterator_range
-void exec_statement(tcp_ssl_connection& conn, const statement& stmt, const std::vector<field>& params)
-{
-    results result;
-    conn.execute(stmt.bind(params.begin(), params.end()), result);
-}
-//]
-
-#ifdef BOOST_ASIO_HAS_CO_AWAIT
-boost::asio::awaitable<void> overview_coro(tcp_ssl_connection& conn)
-{
-    //[overview_async_coroutinescpp20
-    // Using this CompletionToken, you get C++20 coroutines that communicate
-    // errors with error_codes. This way, you can access the diagnostics object.
-    constexpr auto token = boost::asio::as_tuple(boost::asio::use_awaitable);
-
-    // Run our query as a coroutine
-    diagnostics diag;
-    results result;
-    auto [ec] = co_await conn.async_execute("SELECT 'Hello world!'", result, diag, token);
-
-    // This will throw an error_with_diagnostics in case of failure
-    boost::mysql::throw_on_error(ec, diag);
-    //]
-}
-
-#ifndef BOOST_ASIO_USE_TS_EXECUTOR_AS_DEFAULT
-boost::asio::awaitable<void> dont_run()
-{
-    using namespace boost::asio::experimental::awaitable_operators;
-
-    // Setup
-    boost::asio::ssl::context ssl_ctx(boost::asio::ssl::context::tls_client);
-    boost::mysql::tcp_ssl_connection conn(co_await boost::asio::this_coro::executor, ssl_ctx);
-
-    //[overview_async_dont
-    // Coroutine body
-    // DO NOT DO THIS!!!!
-    results result1, result2;
-    co_await (
-        conn.async_execute("SELECT 1", result1, boost::asio::use_awaitable) &&
-        conn.async_execute("SELECT 2", result2, boost::asio::use_awaitable)
-    );
-    //]
-}
-#endif
-#else
-void run_overview_coro(tcp_ssl_connection&) {}
-#endif
-
-void section_overview(tcp_ssl_connection& conn)
-{
-    {
-        //[overview_query_use_case
-        results result;
-        conn.execute("START TRANSACTION", result);
-        //]
-    }
-    {
-        //[overview_statement_use_case
-        statement stmt = conn.prepare_statement(
-            "SELECT first_name FROM employee WHERE company_id = ? AND salary > ?"
-        );
-
-        results result;
-        conn.execute(stmt.bind("HGS", 30000), result);
-        //]
-    }
-    {
-        //[overview_ifaces_table
-        const char* table_definition = R"%(
-            CREATE TEMPORARY TABLE posts (
-                id INT PRIMARY KEY AUTO_INCREMENT,
-                title VARCHAR (256) NOT NULL,
-                body TEXT NOT NULL
-            )
-        )%";
-        //]
-
-        results result;
-        conn.execute(table_definition, result);
-    }
-    {
-        //[overview_ifaces_dynamic
-        // Passing a results object to connection::execute selects the dynamic interface
-        results result;
-        conn.execute("SELECT id, title, body FROM posts", result);
-
-        // Every row is a collection of fields, which are variant-like objects
-        // that represent data. We use as_string() to cast them to the appropriate type
-        for (row_view post : result.rows())
-        {
-            std::cout << "Title: " << post.at(1).as_string() << "Body: " << post.at(2).as_string()
-                      << std::endl;
-        }
-        //]
-    }
-#ifdef BOOST_MYSQL_CXX14
-    {
-        // The struct definition is included above this
-        //[overview_ifaces_static
-        //
-        // This must be placed inside your function or method:
-        //
-
-        // Passing a static_results to execute() selects the static interface
-        static_results<post> result;
-        conn.execute("SELECT id, title, body FROM posts", result);
-
-        // Query results are parsed directly into your own type
-        for (const post& p : result.rows())
-        {
-            std::cout << "Title: " << p.title << "Body: " << p.body << std::endl;
-        }
-        //]
-    }
-#endif
-
-    {
-        //[overview_statements_setup
-        results result;
-        conn.execute(
-            R"%(
-                CREATE TEMPORARY TABLE products (
-                    id VARCHAR(50) PRIMARY KEY,
-                    description VARCHAR(256)
-                )
-            )%",
-            result
-        );
-        conn.execute("INSERT INTO products VALUES ('PTT', 'Potatoes'), ('CAR', 'Carrots')", result);
-        //]
-    }
-    {
-        //[overview_statements_prepare
-        statement stmt = conn.prepare_statement("SELECT description FROM products WHERE id = ?");
-        //]
-
-        //[overview_statements_execute
-        // Obtain the product_id from the user. product_id is untrusted input
-        const char* product_id = get_value_from_user();
-
-        // Execute the statement
-        results result;
-        conn.execute(stmt.bind(product_id), result);
-
-        // Use result as required
-        //]
-
-        conn.execute("DROP TABLE products", result);
-    }
-    {
-        //[overview_errors_sync_errc
-        error_code ec;
-        diagnostics diag;
-        results result;
-
-        // The provided SQL is invalid. The server will return an error.
-        // ec will be set to a non-zero value
-        conn.execute("this is not SQL!", result, ec, diag);
-
-        if (ec)
-        {
-            // The error code will likely report a syntax error
-            std::cout << "Operation failed with error code: " << ec << '\n';
-
-            // diag.server_message() will contain the classic phrase
-            // "You have an error in your SQL syntax; check the manual..."
-            // Bear in mind that server_message() may contain user input, so treat it with caution
-            std::cout << "Server diagnostics: " << diag.server_message() << std::endl;
-        }
-        //]
-    }
-    {
-        //[overview_errors_sync_exc
-        try
-        {
-            // The provided SQL is invalid. This function will throw an exception.
-            results result;
-            conn.execute("this is not SQL!", result);
-        }
-        catch (const error_with_diagnostics& err)
-        {
-            // error_with_diagnostics contains an error_code and a diagnostics object.
-            // It inherits from boost::system::system_error.
-            std::cout << "Operation failed with error code: " << err.code() << '\n'
-                      << "Server diagnostics: " << err.get_diagnostics().server_message() << std::endl;
-        }
-        //]
-    }
-    {
-#ifdef BOOST_ASIO_HAS_CO_AWAIT
-        run_coro(conn.get_executor(), [&conn] { return overview_coro(conn); });
-#endif
-    }
-    {
-        results r;
-        conn.execute("DROP TABLE IF EXISTS posts", r);
-    }
-    {
-        //[overview_multifn
-        // Create the table and some sample data
-        // In a real system, body may be megabaytes long.
-        results result;
-        conn.execute(
-            R"%(
-                CREATE TEMPORARY TABLE posts (
-                    id INT PRIMARY KEY AUTO_INCREMENT,
-                    title VARCHAR (256),
-                    body TEXT
-                )
-            )%",
-            result
-        );
-        conn.execute(
-            R"%(
-                INSERT INTO posts (title, body) VALUES
-                    ('Post 1', 'A very long post body'),
-                    ('Post 2', 'An even longer post body')
-            )%",
-            result
-        );
-
-        // execution_state stores state about our operation, and must be passed to all functions
-        execution_state st;
-
-        // Writes the query request and reads the server response, but not the rows
-        conn.start_execution("SELECT title, body FROM posts", st);
-
-        // Reads all the returned rows, in batches.
-        // st.complete() returns true once there are no more rows to read
-        while (!st.complete())
-        {
-            // row_batch will be valid until conn performs the next network operation
-            rows_view row_batch = conn.read_some_rows(st);
-
-            for (row_view post : row_batch)
-            {
-                // Process post as required
-                std::cout << "Title:" << post.at(0) << std::endl;
-            }
-        }
-        //]
-
-        conn.execute("DROP TABLE posts", result);
-    }
-}
-
-void section_dynamic(tcp_ssl_connection& conn)
-{
-    {
-        //[dynamic_views
-        // Populate a results object
-        results result;
-        conn.execute("SELECT 'Hello world'", result);
-
-        // results::rows() returns a rows_view. The underlying memory is owned by the results object
-        rows_view all_rows = result.rows();
-
-        // Indexing a rows_view yields a row_view. The underlying memory is owned by the results object
-        row_view first_row = all_rows.at(0);
-
-        // Indexing a row_view yields a field_view. The underlying memory is owned by the results object
-        field_view first_field = first_row.at(0);  // Contains the string "Hello world"
-
-        //]
-        ASSERT(first_field.as_string() == "Hello world");
-
-        //[dynamic_taking_ownership
-        // You may use all_rows_owning after result has gone out of scope
-        rows all_rows_owning{all_rows};
-
-        // You may use first_row_owning after result has gone out of scope
-        row first_row_owning{first_row};
-
-        // You may use first_field_owning after result has gone out of scope
-        field first_field_owning{first_field};
-        //]
-    }
-    {
-        //[dynamic_using_fields
-        results result;
-        conn.execute("SELECT 'abc', 42", result);
-
-        // Obtain a field's underlying value using the is_xxx and get_xxx accessors
-        field_view f = result.rows().at(0).at(0);  // f points to the string "abc"
-        if (f.is_string())
-        {
-            // we know it's a string, unchecked access
-            string_view s = f.get_string();
-            std::cout << s << std::endl;  // Use the string as required
-        }
-        else
-        {
-            // Oops, something went wrong - schema msimatch?
-        }
-
-        // Alternative: use the as_xxx accessor
-        f = result.rows().at(0).at(1);
-        std::int64_t value = f.as_int64();  // Checked access. Throws if f doesn't contain an int
-        std::cout << value << std::endl;    // Use the int as required
-
-        //]
-    }
-    {
-        //[dynamic_handling_nulls
-        results result;
-
-        // Create some test data
-        conn.execute(
-            R"%(
-                CREATE TEMPORARY TABLE products (
-                    id VARCHAR(50) PRIMARY KEY,
-                    description VARCHAR(256)
-                )
-            )%",
-            result
-        );
-        conn.execute("INSERT INTO products VALUES ('PTT', 'Potatoes'), ('CAR', NULL)", result);
-
-        // Retrieve the data. Note that some fields are NULL
-        conn.execute("SELECT id, description FROM products", result);
-
-        for (row_view r : result.rows())
-        {
-            field_view description_fv = r.at(1);
-            if (description_fv.is_null())
-            {
-                // Handle the NULL value
-                // Note: description_fv.is_string() will return false here; NULL is represented as a separate
-                // type
-                std::cout << "No description for product_id " << r.at(0) << std::endl;
-            }
-            else
-            {
-                // Handle the non-NULL case. Get the underlying value and use it as you want
-                // If there is any schema mismatch (and description was not defined as VARCHAR), this will
-                // throw
-                string_view description = description_fv.as_string();
-
-                // Use description as required
-                std::cout << "product_id " << r.at(0) << ": " << description << std::endl;
-            }
-        }
-        //]
-
-        conn.execute("DROP TABLE products", result);
-    }
-    {
-        //[dynamic_field_accessor_references
-        field f("my_string");            // constructs a field that owns the string "my_string"
-        std::string& s = f.as_string();  // s points into f's storage
-        s.push_back('2');                // f now holds "my_string2"
-
-        //]
-
-        ASSERT(s == "my_string2");
-    }
-    {
-        //[dynamic_field_assignment
-        field f("my_string");  // constructs a field that owns the string "my_string"
-        f = 42;                // destroys "my_string" and stores the value 42 as an int64
-
-        //]
-
-        ASSERT(f.as_int64() == 42);
-    }
-}
-
-void section_static(tcp_ssl_connection& conn)
-{
-    boost::ignore_unused(conn);
-#ifdef BOOST_MYSQL_CXX14
-    {
-        //[static_setup
-        const char* table_definition = R"%(
-            CREATE TEMPORARY TABLE posts (
-                id INT PRIMARY KEY AUTO_INCREMENT,
-                title VARCHAR (256) NOT NULL,
-                body TEXT NOT NULL
-            )
-        )%";
-        const char* query = "SELECT id, title, body FROM posts";
-        //]
-
-        results r;
-        conn.execute(table_definition, r);
-
-        //[static_query
-        static_results<post> result;
-        conn.execute(query, result);
-
-        for (const post& p : result.rows())
-        {
-            // Process the post as required
-            std::cout << "Title: " << p.title << "\n" << p.body << "\n";
-        }
-        //]
-
-        conn.execute("DROP TABLE posts", r);
-    }
-    {
-        //[static_field_order
-        // Summing 0e0 is MySQL way to cast a DECIMAL field to DOUBLE
-        const char* sql = R"%(
-            SELECT
-                IFNULL(AVG(salary), 0.0) + 0e0 AS average,
-                IFNULL(MAX(salary), 0.0) + 0e0 AS max_value,
-                company_id AS company
-            FROM employee
-            GROUP BY company_id
-        )%";
-
-        static_results<statistics> result;
-        conn.execute(sql, result);
-        //]
-    }
-    {
-        //[static_tuples
-        static_results<std::tuple<std::int64_t>> result;
-        conn.execute("SELECT COUNT(*) FROM employee", result);
-        std::cout << "Number of employees: " << std::get<0>(result.rows()[0]) << "\n";
-        //]
-    }
-#ifndef BOOST_NO_CXX17_HDR_OPTIONAL
-    {
-        //[static_nulls_table
-        const char* table_definition = R"%(
-            CREATE TEMPORARY TABLE posts_v2 (
-                id INT PRIMARY KEY AUTO_INCREMENT,
-                title VARCHAR (256) NOT NULL,
-                body TEXT
-            )
-        )%";
-        //]
-
-        // Verify that post_v2's definition is correct
-        results r;
-        conn.execute(table_definition, r);
-        static_results<post_v2> result;
-        conn.execute("SELECT * FROM posts_v2", result);
-        conn.execute("DROP TABLE posts_v2", r);
-    }
-#endif  // BOOST_NO_CXX17_HDR_OPTIONAL
-#endif  // BOOST_MYSQL_CXX14
-}
-
-void section_prepared_statements(tcp_ssl_connection& conn)
-{
-    {
-        //[prepared_statements_prepare
-        // Table setup
-        const char* table_definition = R"%(
-            CREATE TEMPORARY TABLE products (
-                id INT PRIMARY KEY AUTO_INCREMENT,
-                description VARCHAR(256),
-                price INT NOT NULL,
-                show_in_store TINYINT
-            )
-        )%";
-        results result;
-        conn.execute(table_definition, result);
-
-        // Prepare a statement to insert into this table
-        statement stmt = conn.prepare_statement(
-            "INSERT INTO products (description, price, show_in_store) VALUES (?, ?, ?)"
-        );
-        //]
-
-        // Run the functions to verify that evth works
-        insert_product(conn, stmt, string_view("This is a product"), 2000, true);
-        run_insert_product_optional(conn, stmt);
-        exec_statement(conn, stmt, {field_view("abc"), field_view(2000), field_view(1)});
-        conn.execute("DROP TABLE products", result);
-    }
-    {
-        //[prepared_statements_casting_table
-        const char* table_definition = "CREATE TEMPORARY TABLE my_table(my_field TINYINT)";
-        //]
-
-        results r;
-        conn.execute(table_definition, r);
-
-        //[prepared_statements_casting_execute
-        int value = get_int_value_from_user();
-        auto stmt = conn.prepare_statement("INSERT INTO my_table VALUES (?)");
-
-        results result;
-        conn.execute(stmt.bind(value), result);
-        //]
-    }
-}
 
 void section_multi_resultset(tcp_ssl_connection& conn)
 {
@@ -2286,10 +1728,6 @@ void main_impl()
     conn.connect(*endpoints.begin(), params);
     //]
 
-    section_overview(conn);
-    section_dynamic(conn);
-    section_static(conn);
-    section_prepared_statements(conn);
     section_multi_resultset(conn);
     section_multi_resultset_multi_queries();
     section_multi_function(conn);
